onestop-python-client/onestop/schemas/psiSchemaClasses_test/org/cedar/schemas/avro/geojson/line_string_factory.py
from factory import Factory, lazy_attribute

from psiSchemaClasses.org.cedar.schemas.avro.geojson import LineString


class LineStringFactory(Factory):
    class Meta:
        model = LineString

        pass
