onestop-python-client/requirements.txt
confluent-kafka
avro-python3
fastavro
smart-open
PyYAML~=5.3.1
setuptools~=49.2.0
argparse~=1.4.0
boto3~=1.15.11
requests~=2.24.0
botocore~=1.18.11
moto==1.3.16.dev122
undictify
